Move badges to repository match header section
