+++ client/branded/src/search-ui/components/ResultContainer.module.scss
@@ -59,18 +59,3 @@
 .highlight-result.result:focus-within {
     box-shadow: 0 0 0 2px var(--border-color-2);
 }
-
-.repo-metadata {
-    gap: 0.5rem;
-    font-size: 0.75rem;
-
-    &-key {
-        border-top-right-radius: 0;
-        border-bottom-right-radius: 0;
-    }
-
-    &-value {
-        border-top-left-radius: 0;
-        border-bottom-left-radius: 0;
-    }
-}
